# backend/openapi.json
{
  "components": {
    "schemas": {
      "AimTimeline": {
        "properties": {
          "default_aim": {
            "discriminator": {
              "mapping": {
                "follow_path": "#/components/schemas/FollowPathAim",
                "look_at_point": "#/components/schemas/LookAtPointAim"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/FollowPathAim"
              },
              {
                "$ref": "#/components/schemas/LookAtPointAim"
              }
            ],
            "title": "Default Aim"
          },
          "keyframes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CameraKeyframe"
            },
            "title": "Keyframes",
            "type": "object"
          },
          "world_up": {
            "default": [
              0.0,
              1.0,
              0.0
            ],
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "World Up",
            "type": "array"
          }
        },
        "title": "AimTimeline",
        "type": "object"
      },
...
      "CameraTrackUpdate": {
        "properties": {
          "default_aim": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "follow_path": "#/components/schemas/FollowPathAim",
                    "look_at_point": "#/components/schemas/LookAtPointAim"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/FollowPathAim"
                  },
                  {
                    "$ref": "#/components/schemas/LookAtPointAim"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Aim"
          },
          "world_up": {
            "anyOf": [
              {
                "maxItems": 3,
                "minItems": 3,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "World Up"
          }
        },
        "title": "CameraTrackUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Create camera-path projects, author scene and trajectory data, edit playback timelines, and run the trajectory agent. Canonical operations use project revision ETags for optimistic concurrency.",
    "summary": "Semantic 3D camera trajectory authoring API",
    "title": "Camera Path API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/camera": {
      "get": {
        "description": "Return camera aim defaults and keyframes.",
        "operationId": "getAimTimeline",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimTimeline"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the aim timeline",
        "tags": [
          "Timelines"
        ]
      },
      "patch": {
        "description": "Update aim defaults and return the changed timeline.",
        "operationId": "updateCameraTrack",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraTrackUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimTimeline"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update the aim timeline",
        "tags": [
          "Timelines"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/camera/depth-of-field/keyframes": {
      "post": {
        "description": "Create and return a depth-of-field keyframe.",
        "operationId": "createDepthOfFieldKeyframe",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepthOfFieldKeyframeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepthOfFieldKeyframe"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a depth-of-field keyframe",
        "tags": [
          "Timelines"
        ]
      }
    },
    "/api/v1/projects/{project_id}/camera/depth-of-field/keyframes/{keyframe_id}": {
      "delete": {
        "description": "Delete a depth-of-field keyframe.",
        "operationId": "deleteDepthOfFieldKeyframe",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "keyframe_id",
            "required": true,
            "schema": {
              "title": "Keyframe Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found.",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a depth-of-field keyframe",
        "tags": [
          "Timelines"
        ]
      },
      "patch": {
        "description": "Update and return a depth-of-field keyframe.",
        "operationId": "updateDepthOfFieldKeyframe",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "keyframe_id",
            "required": true,
            "schema": {
              "title": "Keyframe Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepthOfFieldKeyframeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepthOfFieldKeyframe"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a depth-of-field keyframe",
        "tags": [
          "Timelines"
        ]
      }
    },
    "/api/v1/projects/{project_id}/camera/keyframes": {
      "post": {
        "description": "Create and return a camera aim keyframe.",
        "operationId": "createCameraAimKeyframe",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraKeyframeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CameraKeyframe"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create an aim keyframe",
        "tags": [
          "Timelines"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/camera/orientation": {
      "get": {
        "description": "Return default camera orientation and orientation keyframes.",
        "operationId": "getOrientationTimeline",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrientationTimeline"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the orientation timeline",
        "tags": [
          "Timelines"
        ]
      },
      "patch": {
        "description": "Update default orientation and return the changed timeline.",
        "operationId": "updateCameraOrientation",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "examples": [
                "\"4\""
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraOrientation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrientationTimeline"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Revision of the returned project snapshot.",
                "schema": {
                  "examples": [
                    "\"4\""
                  ],
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Valid editor JWT required."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Revision or reference conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or trajectory geometry."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The If-Match header is required."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update default orientation",
        "tags": [
          "Timelines"
        ]
      }
    },
...
  },
...
}